Big refactor of camera code to minimise duplication
