+++ pocs/camera/canon_gphoto2.py
@@ -86,7 +86,11 @@ class Camera(AbstractGPhotoCamera):
         # To be used for marking when exposure is complete (see `process_exposure`)
         camera_event = Event()
 
-        exp_time, file_path, metadata = self._setup_observation(observation, filename, **kwargs)
+        exp_time, file_path, image_id, metadata = self._setup_observation(observation,
+                                                                          headers,
+                                                                          filename,
+                                                                          *args,
+                                                                          **kwargs)
 
         proc = self.take_exposure(seconds=exp_time, filename=file_path)
 
@@ -144,52 +148,11 @@ class Camera(AbstractGPhotoCamera):
         else:
             return proc
 
-    def process_exposure(self, info, signal_event, exposure_process=None):
-        """Processes the exposure
-
-        Converts the CR2 to a FITS file. If the camera is a primary camera, extract the
-        jpeg image and save metadata to mongo `current` collection. Saves metadata
-        to mongo `observations` collection for all images
-
-        Args:
-            info (dict): Header metadata saved for the image
-            signal_event (threading.Event): An event that is set signifying that the
-                camera is done with this exposure
+    def _process_fits(self, filepath, info):
         """
-        if exposure_process:
-            exposure_process.wait()
-
-        image_id = info['image_id']
-        seq_id = info['sequence_id']
-        file_path = info['file_path']
-        self.logger.debug("Processing {}".format(image_id))
-
-        try:
-            self.logger.debug("Extracting pretty image")
-            img_utils.make_pretty_image(file_path, title=image_id, primary=info['is_primary'])
-        except Exception as e:
-            self.logger.warning('Problem with extracting pretty image: {}'.format(e))
-
+        Converts the CR2 to a FITS file
+        """
         self.logger.debug("Converting CR2 -> FITS: {}".format(file_path))
         fits_path = cr2_utils.cr2_to_fits(file_path, headers=info, remove_cr2=True)
-
         # Replace the path name with the FITS file
         info['file_path'] = fits_path
-
-        if info['is_primary']:
-            self.logger.debug("Adding current observation to db: {}".format(image_id))
-            self.db.insert_current('observations', info, include_collection=False)
-        else:
-            self.logger.debug('Compressing {}'.format(file_path))
-            fits_utils.fpack(fits_path)
-
-        self.logger.debug("Adding image metadata to db: {}".format(image_id))
-        self.db.observations.insert_one({
-            'data': info,
-            'date': current_time(datetime=True),
-            'type': 'observations',
-            'sequence_id': seq_id,
-        })
-
-        # Mark the event as done
-        signal_event.set()
