+++ pocs/camera/sbig.py
@@ -1,16 +1,11 @@
 from threading import Event
-from threading import Thread
 
 from astropy import units as u
 from astropy.io import fits
 
-from pocs.utils import current_time
-from pocs.utils import images
 from pocs.camera import AbstractCamera
 from pocs.camera.sbigudrv import INVALID_HANDLE_VALUE
 from pocs.camera.sbigudrv import SBIGDriver
-from pocs.focuser.birger import Focuser as BirgerFocuser
-
 
 class Camera(AbstractCamera):
 
@@ -99,6 +94,7 @@ class Camera(AbstractCamera):
         self.logger.debug("{} connected".format(self.name))
         self._connected = True
         self._serial_number = self._info['serial number']
+        self.model = self._info['camera name']
 
         if self._info['colour']:
             if self._info['Truesense']:
@@ -107,38 +103,6 @@ class Camera(AbstractCamera):
                 self.filter_type = 'RGGB'
         else:
             self.filter_type = 'M'
-
-    def take_observation(self, observation, headers=None, filename=None, *args, **kwargs):
-        """Take an observation
-
-        Gathers various header information, sets the file path, and calls
-            `take_exposure`. Also creates a `threading.Event` object and a
-            `threading.Thread` object. The Thread calls `process_exposure`
-            after the exposure had completed and the Event is set once
-            `process_exposure` finishes.
-
-        Args:
-            observation (~pocs.scheduler.observation.Observation): Object
-                describing the observation
-            headers (dict): Header data to be saved along with the file.
-            **kwargs (dict): Optional keyword arguments (`exp_time`, dark)
-
-        Returns:
-            threading.Event: An event to be set when the image is done processing
-        """
-        # To be used for marking when exposure is complete (see `process_exposure`)
-        camera_event = Event()
-
-        exp_time, file_path, metadata = self._setup_observation(observation, filename, **kwargs)
-
-        exposure_event = self.take_exposure(seconds=exp_time, filename=file_path, **kwargs)
-
-        # Process the exposure once readout is complete
-        t = Thread(target=self.process_exposure, args=(metadata, camera_event, exposure_event))
-        t.name = '{}Thread'.format(self.name)
-        t.start()
-
-        return camera_event
 
     def take_exposure(self,
                       seconds=1.0 * u.second,
@@ -176,70 +140,6 @@ class Camera(AbstractCamera):
 
         return exposure_event
 
-    def process_exposure(self, info, signal_event, exposure_event=None):
-        """
-        Processes the exposure
-
-        Args:
-            info (dict): Header metadata saved for the image
-            signal_event (threading.Event): An event that is set signifying that the
-                camera is done with this exposure
-            exposure_event (threading.Event, optional): An event that should be set
-                when the exposure is complete, triggering the processing.
-        """
-        # If passed an Event that signals the end of the exposure wait for it to be set
-        if exposure_event:
-            exposure_event.wait()
-
-        image_id = info['image_id']
-        file_path = info['file_path']
-        self.logger.debug("Processing {}".format(image_id))
-
-        # Add FITS headers from info the same as images.cr2_to_fits()
-        self.logger.debug("Updating FITS headers: {}".format(file_path))
-        with fits.open(file_path, 'update') as f:
-            hdu = f[0]
-            hdu.header.set('IMAGEID', info.get('image_id', ''))
-            hdu.header.set('SEQID', info.get('sequence_id', ''))
-            hdu.header.set('FIELD', info.get('field_name', ''))
-            hdu.header.set('RA-MNT', info.get('ra_mnt', ''), 'Degrees')
-            hdu.header.set('HA-MNT', info.get('ha_mnt', ''), 'Degrees')
-            hdu.header.set('DEC-MNT', info.get('dec_mnt', ''), 'Degrees')
-            hdu.header.set('EQUINOX', info.get('equinox', 2000.))  # Assume J2000
-            hdu.header.set('AIRMASS', info.get('airmass', ''), 'Sec(z)')
-            hdu.header.set('FILTER', info.get('filter', ''))
-            hdu.header.set('LAT-OBS', info.get('latitude', ''), 'Degrees')
-            hdu.header.set('LONG-OBS', info.get('longitude', ''), 'Degrees')
-            hdu.header.set('ELEV-OBS', info.get('elevation', ''), 'Meters')
-            hdu.header.set('MOONSEP', info.get('moon_separation', ''), 'Degrees')
-            hdu.header.set('MOONFRAC', info.get('moon_fraction', ''))
-            hdu.header.set('CREATOR', info.get('creator', ''), 'POCS Software version')
-            hdu.header.set('INSTRUME', info.get('camera_uid', ''), 'Camera ID')
-            hdu.header.set('OBSERVER', info.get('observer', ''), 'PANOPTES Unit ID')
-            hdu.header.set('ORIGIN', info.get('origin', ''))
-            hdu.header.set('RA-RATE', info.get('tracking_rate_ra', ''), 'RA Tracking Rate')
-
-        if info['is_primary']:
-            self.logger.debug("Extracting pretty image")
-            images.make_pretty_image(file_path, title=info['field_name'], primary=True)
-
-            self.logger.debug("Adding current observation to db: {}".format(image_id))
-            self.db.insert_current('observations', info, include_collection=False)
-        else:
-            self.logger.debug('Compressing {}'.format(file_path))
-            images.fpack(file_path)
-
-        self.logger.debug("Adding image metadata to db: {}".format(image_id))
-        self.db.observations.insert_one({
-            'data': info,
-            'date': current_time(datetime=True),
-            'type': 'observations',
-            'image_id': image_id,
-        })
-
-        # Mark the event as done
-        signal_event.set()
-
 # Private methods
 
     def _fits_header(self, seconds, dark):
@@ -248,7 +148,6 @@ class Camera(AbstractCamera):
         # Unbinned. Need to chance if binning gets implemented.
         readout_mode = 'RM_1X1'
 
-        header.set('CAM-MOD', self._info['camera name'], 'Camera model')
         header.set('CAM-FW', self._info['firmware version'], 'Camera firmware version')
         header.set('XPIXSZ', self._info['readout modes'][readout_mode]['pixel width'].value,
                    'Microns')
